Revert `uiData` -> `index` moves
